# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.createModelRequest:
    properties:
      aliases:
        description: Aliases are optional short names that resolve to this model in
          proxy requests.
        items:
          type: string
        type: array
      api_key:
        type: string
      azure_api_version:
        type: string
      azure_deployment:
        type: string
      base_url:
        type: string
      bill_min_per_request:
        type: boolean
      bill_per_request:
        type: boolean
      bill_per_token:
        type: boolean
      fallback_model_name:
        description: |-
          FallbackModelName is the name of the model to try if all deployments of
          this model are unavailable. Requires an Enterprise license with the
          fallback_chains feature. Leave empty to disable fallback.
        type: string
      gcp_location:
        description: GCPLocation is the Google Cloud region (e.g. "us-central1").
          Required when provider is "vertex".
        type: string
      gcp_project:
        description: GCPProject is the Google Cloud project ID. Required when provider
          is "vertex".
        type: string
      input_price_per_1m:
        type: number
      is_public:
        description: IsPublic marks the model as visible on the public storefront
          price list.
        type: boolean
      logo:
        description: Logo is the customer-facing logo URL or asset path.
        type: string
      max_context_tokens:
        type: integer
      max_retries:
        description: |-
          MaxRetries is the maximum number of deployments to attempt before returning
          an error. 0 means try all available deployments.
        type: integer
      name:
        type: string
      output_price_per_1m:
        type: number
      pii_filter:
        description: |-
          PIIFilter explicitly enables or disables PII anonymization for requests
          routed to this model. When omitted the network-level default is used.
          Pass true to force anonymization on, false to force it off.
        type: boolean
      provider:
        type: string
      rpm_limit:
        type: integer
      sell_cached_input_per_1m:
        type: number
      sell_input_per_1m:
        description: |-
          SellInputPer1M / SellOutputPer1M / SellCachedInputPer1M are the
          customer-facing prices in USD per 1M tokens. Omit to leave the model
          unbilled (requests do not debit wallets).
        type: number
      sell_min_per_request:
        type: number
      sell_output_per_1m:
        type: number
      sell_per_request:
        type: number
      strategy:
        description: |-
          Strategy is the load balancing strategy used when multiple deployments are
          configured. Valid values: "round-robin", "weighted", "priority". Omit or
          pass an empty string for single-deployment (legacy) mode.
        type: string
      timeout:
        description: |-
          Timeout is the per-model upstream timeout as a Go duration string (e.g. "30s",
          "2m"). When non-empty it overrides the global stream/response timeout for
          this model. Omit or pass an empty string to use the global default.
        type: string
      type:
        type: string
    type: object
  admin.deploymentResponse:
    properties:
      azure_api_version:
        type: string
      azure_deployment:
        type: string
      base_url:
        type: string
      cost_cache_write_per_1m:
        type: number
      cost_cached_input_per_1m:
        type: number
      cost_input_per_1m:
        type: number
      cost_output_per_1m:
        type: number
      created_at:
        type: string
      daily_request_limit:
        type: integer
      gcp_location:
        description: GCPLocation is the Google Cloud region. Non-empty only for provider
          "vertex".
        type: string
      gcp_project:
        description: GCPProject is the Google Cloud project ID. Non-empty only for
          provider "vertex".
        type: string
      id:
        type: string
      is_active:
        type: boolean
      model_id:
        type: string
      name:
        type: string
      pii_filter:
        description: |-
          PIIFilter is the per-deployment PII anonymization override. Nil means
          inherit from the model-level or network-level default; true forces
          anonymization on; false forces it off.
        type: boolean
      priority:
        type: integer
      provider:
        type: string
      provider_id:
        type: string
      rpm_limit:
        type: integer
      tpm_limit:
        type: integer
      updated_at:
        type: string
      upstream_model:
        description: |-
          UpstreamModel is the model name sent upstream; empty means the
          canonical model name is used.
        type: string
      weight:
        type: integer
    type: object
...
  admin.modelResponse:
    properties:
      aliases:
        items:
          type: string
        type: array
      azure_api_version:
        type: string
      azure_deployment:
        type: string
      base_url:
        type: string
      bill_min_per_request:
        type: boolean
      bill_per_request:
        type: boolean
      bill_per_token:
        type: boolean
      created_at:
        type: string
      deployments:
        description: Deployments contains the model's deployment entries when present.
        items:
          $ref: '#/definitions/admin.deploymentResponse'
        type: array
      fallback_model_name:
        description: |-
          FallbackModelName is the name of the fallback model, or empty when none
          is configured.
        type: string
      gcp_location:
        description: GCPLocation is the Google Cloud region. Non-empty only for provider
          "vertex".
        type: string
      gcp_project:
        description: GCPProject is the Google Cloud project ID. Non-empty only for
          provider "vertex".
        type: string
      id:
        type: string
      input_price_per_1m:
        type: number
      is_active:
        type: boolean
      is_public:
        description: IsPublic marks the model as visible on the public storefront
          price list.
        type: boolean
      logo:
        description: Logo is the customer-facing logo URL or asset path.
        type: string
      max_context_tokens:
        type: integer
      max_retries:
        description: |-
          MaxRetries is the maximum number of deployments to attempt before
          returning an error. 0 means try all available deployments.
        type: integer
      name:
        type: string
      output_price_per_1m:
        type: number
      pii_filter:
        description: |-
          PIIFilter is the per-model PII anonymization override. Nil means the
          network-level default is used; true forces anonymization on; false off.
        type: boolean
      provider:
        type: string
      route_count:
        description: RouteCount is the number of enabled combo route steps (API products
          only).
        type: integer
      routing_strategy:
        type: string
      rpm_limit:
        type: integer
      sell_cached_input_per_1m:
        type: number
      sell_input_per_1m:
        description: Sell prices are the customer-facing prices per 1M tokens.
        type: number
      sell_min_per_request:
        type: number
      sell_output_per_1m:
        type: number
      sell_per_request:
        type: number
      source:
        type: string
      sticky_round_robin_limit:
        type: integer
      strategy:
        description: |-
          Strategy is the load balancing strategy used when multiple deployments
          are configured (e.g. "round-robin", "weighted", "priority"). An empty
          string means single-deployment (legacy) mode.
        type: string
      timeout:
        description: |-
          Timeout is the per-model upstream timeout (e.g. "30s", "2m").
          An empty string means the global default is used.
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  admin.paginatedModelsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/admin.modelResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
    type: object
...
  admin.swaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierror.SwaggerDetail'
    type: object
...
  apierror.SwaggerDetail:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: resource not found
        type: string
    type: object
info:
  contact: {}
  description: Prepaid LLM API marketplace with OpenAI-compatible proxy, wallet billing,
    and model catalog.
  title: Tavo API
  version: 0.2.0
paths:
...
  /models:
    get:
      description: Returns a cursor-paginated list of models. Requires system admin.
      parameters:
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Pagination cursor (UUIDv7 of the last seen model)
        in: query
        name: cursor
        type: string
      - description: Include inactive models
        in: query
        name: include_inactive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.paginatedModelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List models
      tags:
      - models
    post:
      consumes:
      - application/json
      description: Persists a new model and adds it to the live registry. The API
        key is encrypted at rest. Requires system admin.
      parameters:
      - description: Model parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.createModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin.modelResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a model
      tags:
      - models
...
securityDefinitions:
  BearerAuth:
    description: API key (vl_uk_) or session key (vl_sk_)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...